fix: rename paths- and titles- files
