+++ src/components/Header/HeaderNav.tsx
@@ -3,7 +3,7 @@ import { Typography, createTheme } from '@mui/material';
 import { themeOptions } from '../../assets/theme';
 import { PAGES_HEADER_MENU } from '../../data/pages';
 import { Link } from 'react-router-dom';
-import { PATH } from '../../data/pathsTemp';
+import { PATH } from '../../data/paths';
 
 const plantsTheme = createTheme(themeOptions);
