+++ src/main.css
@@ -8,7 +8,7 @@ a {
   transition: all 0.2s linear;
 }
 
-.link:hover{
+.link:hover {
   color: #fbde00d0;
 }
