+++ src/pages/Guarantee/GuaranteePage.tsx
@@ -1,5 +1,5 @@
 import { Box, useTheme, Typography } from '@mui/material';
-import { PAGES_TITLES } from '../../data/titlesTemp';
+import { PAGES_TITLES } from '../../data/titles';
 import { TEXT_GUARANTEE } from './dataGuarantee';
 import GuaranteeImage from './GuaranteeImg';
