+++ src/pages/Login/LoginPage.tsx
@@ -2,8 +2,8 @@ import { Box, useTheme, Typography, Button } from '@mui/material';
 
 import LoginForm from '../../components/loginForm/LoginForm';
 import { Link } from 'react-router-dom';
-import { PAGES_TITLES } from '../../data/titlesTemp';
-import { PATH } from '../../data/pathsTemp';
+import { PAGES_TITLES } from '../../data/titles';
+import { PATH } from '../../data/paths';
 
 export function LoginPage() {
   const plantsTheme = useTheme();
