+++ src/pages/Registration/RegistrationPage.tsx
@@ -1,8 +1,8 @@
 import Form1 from '../../components/registrationForm/RegistrationForm';
 import { Box, useTheme, Button, Typography } from '@mui/material';
 import { Link } from 'react-router-dom';
-import { PATH } from '../../data/pathsTemp';
-import { PAGES_TITLES } from '../../data/titlesTemp';
+import { PATH } from '../../data/paths';
+import { PAGES_TITLES } from '../../data/titles';
 
 export function RegistrationPage() {
   const plantsTheme = useTheme();
